refactor: drop the preserved-commit push-retry machinery; keep the narrow never-fatal correction outcome
The review rounds on this PR grew a publish-dependent retry mechanism for
commits preserved by a needs_human correction outcome (markers, retry-head
exemptions, resync suppression across the fix cycle, CI repair and SyncBase).
Each round surfaced a new P1 in that mechanism (non-fast-forward loops,
SyncBase hard-resets, stale markers after CI pushes). Development meanwhile
took the stricter rule from PR #929: path membership alone is never FIXED
evidence.

Restore fix_cycle.py, helpers.py, ci_ops.py and remote_repair_unpublished.py to
the development versions, delete fix_cycle_preserved_commit.py and its tests,
and keep only the verdict-protocol change on top of the new base:

- a correction-attempt FIXED whose contentful commit carries no item-scoped
  evidence is preserved and escalated to needs_human instead of terminating
  the protocol (the shape that failed ws_46bc0f45 and ws_8082fef3);
- the contentful probe sits under the same rollback handling as the other
  evidence probes, and the escalation cites the post-sink HEAD it preserves;
- the self-citation guard applies on every correction attempt.

On a push failure the preserved commit follows the ordinary unpublished-repair
path; the recorded needs_human reason keeps describing the change.
