+++ tests/unit/runtime/test_pr_monitor_pre_push_validation_fix_pass_parts/test_pr_monitor_pre_push_validation_fix_pass_part_011.py
@@ -783,5 +783,4 @@ async def test_malformed_first_attempt_with_off_anchor_commit_is_preserved_not_a
     )
 
     assert result.verdict == "needs_human"
-    assert result.preserved_unpublished_commit is True
     assert _git(worktree, "rev-parse", "HEAD").stdout.strip() == edited_head
