+++ tests/unit/runtime/test_pr_monitor_verdict_retry_parts/test_pr_monitor_verdict_retry_part_008.py
@@ -368,7 +368,6 @@ def test_correction_self_citation_outcome_preserves_explicit_needs_human() -> No
     assert outcome.verdict == "needs_human"
     # The commit stays in the worktree, so the escalation is publish-dependent
     # exactly like the no-evidence one (PRRT_kwDOSJAM6s6fpjBw).
-    assert outcome.preserved_unpublished_commit is True
     assert outcome.reason is not None
     assert "human review" in outcome.reason.lower() or "Agent reason:" in outcome.reason
 
@@ -478,7 +477,6 @@ def test_correction_self_citation_outcome_bounds_the_stored_reason() -> None:
     )
     assert fixed.verdict == "fix_committed"
     # An accepted fix already flows through the ordinary publish-dependent path.
-    assert fixed.preserved_unpublished_commit is False
     assert fixed.reason is not None
     assert len(fixed.reason) == 500
     assert fixed.reason.endswith("…")
@@ -491,7 +489,6 @@ def test_correction_self_citation_outcome_bounds_the_stored_reason() -> None:
         has_path_evidence=False,
     )
     assert escalated.verdict == "needs_human"
-    assert escalated.preserved_unpublished_commit is True
     assert escalated.reason is not None
     assert escalated.reason.endswith("Agent reason: short")
